transverse: call appState.openDirectory when opening a directory from anywhere in the app
This allows to have a single point of failure and to show the same error message.

Also fixed the opening of a folder in the inactive view in all cases.
